+++ test/github-issues/1600/entity/User.ts
@@ -1,0 +1,12 @@
+import {Column, Entity, PrimaryGeneratedColumn} from "../../../../src";
+
+@Entity()
+export class User {
+
+    @PrimaryGeneratedColumn()
+    id: number;
+
+    @Column("varchar", { array: true })
+    names: string[];
+
+}